refactor: IdPay prefix for barcode feature
