+++ ts/features/idpay/barcode/components/IdPayBarcodeExpireProgressBar.tsx
